MAT-7405 refactor cql definition builder UI
